Moved ProcessApiRequest tests into class following agreed convention
